+++ docs/source/changelog.rst
@@ -16,6 +16,7 @@ Version (dev)
 * Add support for parsing Strata's Keyvalues-based `~srctools.cmdseq` variant.
 * Allow customising the filename used for injected `~srctools.packlist.PackList` files.
 * Add support for parsing Strata's tintable overlays BSP format.
+* Rearrange `srctools.bsp.Overlay` attributes, parse basis U/V out of the 4 handle positions.
 
 -------------
 Version 2.6.2